requirements.txt
discord.py
openai
apscheduler
pytz
pymongo
flask
python-dotenv